Float 자료형을 모두 Double로 변경

+++ src/AddOperation.kt
@@ -1,5 +1,5 @@
-class AddOperation(num1:Float,num2:Float) : Calculator(num1,num2) {
-    var resultAdd : Float = num1+num2
+class AddOperation(num1:Double,num2:Double) : Calculator(num1,num2) {
+    var resultAdd : Double = num1+num2
 
     init {
         this.resultAdd = resultAdd
